requirements.txt
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
langchain-chroma==0.1.4
langchain-text-splitters==0.3.0
chromadb==0.5.0
pypdf==4.0.1
streamlit==1.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
sentence-transformers==2.3.1    